@@ -2,8 +2,8 @@ apply plugin: 'java'
 apply plugin: 'distribution'
 
 compileJava.options.encoding = 'UTF-8'
-sourceCompatibility = 1.6
-targetCompatibility = 1.6
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
