@@ -31,6 +31,7 @@ package jpass.ui;
 import java.awt.BorderLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.function.Consumer;
 
 import javax.swing.AbstractAction;
 import javax.swing.Action;
@@ -43,11 +44,8 @@ import javax.swing.border.EmptyBorder;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
 
-import jpass.ui.action.Callback;
-
 import static javax.swing.KeyStroke.getKeyStroke;
 import static java.awt.event.KeyEvent.VK_ESCAPE;
-import static javax.swing.JComponent.WHEN_IN_FOCUSED_WINDOW;
 
 /**
  * Class for representing search panel. Search panel is hidden by default.
@@ -65,14 +63,14 @@ public class SearchPanel extends JPanel implements ActionListener {
     private final JLabel label;
     private final JTextField criteriaField;
     private final JButton closeButton;
-    private final Callback callback;
+    private final transient Consumer<Boolean> callback;
 
     /**
      * Creates a new search panel with the given callback object.
      *
      * @param searchCallback the callback used on document updates.
      */
-    public SearchPanel(Callback searchCallback) {
+    public SearchPanel(Consumer<Boolean> searchCallback) {
         super(new BorderLayout());
         setBorder(new EmptyBorder(2, 2, 2, 2));
 
@@ -86,17 +84,17 @@ public class SearchPanel extends JPanel implements ActionListener {
             this.criteriaField.getDocument().addDocumentListener(new DocumentListener() {
                 @Override
                 public void changedUpdate(DocumentEvent e) {
-                    callback.call(isEnabled());
+                    callback.accept(isEnabled());
                 }
 
                 @Override
                 public void insertUpdate(DocumentEvent e) {
-                    callback.call(isEnabled());
+                    callback.accept(isEnabled());
                 }
 
                 @Override
                 public void removeUpdate(DocumentEvent e) {
-                    callback.call(isEnabled());
+                    callback.accept(isEnabled());
                 }
             });
         }
