@@ -28,6 +28,7 @@
  */
 package jpass.util;
 
+import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.SecureRandom;
 import java.util.Random;
@@ -88,7 +89,7 @@ public final class CryptUtils {
         MessageDigest md = MessageDigest.getInstance("SHA-256");
         md.reset();
         // md.update(salt);
-        byte[] bytes = new String(text).getBytes("UTF-8");
+        byte[] bytes = new String(text).getBytes(StandardCharsets.UTF_8);
         byte[] digest = md.digest(bytes);
         for (int i = 0; i < iteration; i++) {
             md.reset();
